Fix the pb during assignment of multilevel jobs
For multilevel job the dispatcher only assigns N command for each cycle, N being the number of children of the job (when it's a folder node)
This is due to the dispatchIterator function that yields assignment tuples. It was returning prematurly after iterating over a node's children instead of iterating until maxRN or queue size reached.