HACK to update katana license number
NOTE: For Mikros-DDD only
A pb with katana license checking and RLM causes overuse of katana
license token. The license counter in Puli does not match the RLM counter
anymore.
To allow better use of license until the end of the project, the number
of used license is calculated with rlmutil and stored to puli's
counter value.